more plots updates
